chore: move rust server to websockets
