+++ .vscode/extensions.json
@@ -1,0 +1,7 @@
+{
+  "recommendations": [
+    "rust-lang.rust-analyzer",
+    "esbenp.prettier-vscode",
+    "ms-python.python"
+  ]
+}
