--- .vscode/settings.json
@@ -1,7 +1,0 @@
-{
-    "rust-analyzer.linkedProjects": [
-        "./Cargo.toml",
-        "./Cargo.toml",
-        "./Cargo.toml"
-    ]
-}
