--- src-ws-server/main.rs
@@ -1,48 +1,0 @@
-pub mod diff;
-pub mod utils;
-pub mod handlers;
-
-use std::net::{TcpListener, TcpStream};
-use std::thread::spawn;
-use tungstenite::{accept, handshake::HandshakeRole, Error, HandshakeError, Message, Result};
-
-use crate::handlers::Cmd;
-
-fn must_not_block<Role: HandshakeRole>(err: HandshakeError<Role>) -> Error {
-    match err {
-        HandshakeError::Interrupted(_) => panic!("Bug: blocking socket would block"),
-        HandshakeError::Failure(f) => f,
-    }
-}
-
-fn handle_client(stream: TcpStream) -> Result<()> {
-    let mut socket = accept(stream).map_err(must_not_block)?;
-    loop {
-        match socket.read()? {
-            msg @ Message::Text(_) | msg @ Message::Binary(_) => {
-                socket.send(match Cmd::handle(msg.to_string()) {
-                    Ok(res) => res,
-                    Err(e) => Message::Text(e.to_string())
-                })?;
-            }
-            Message::Ping(_) | Message::Pong(_) | Message::Close(_) | Message::Frame(_) => {}
-        }
-    }
-}
-
-fn main () {
-    let server = TcpListener::bind("127.0.0.1:8000").unwrap();
-    for stream in server.incoming() {
-        spawn(move || match stream {
-            Ok(stream) => {
-                if let Err(err) = handle_client(stream) {
-                    match err {
-                        Error::ConnectionClosed | Error::Protocol(_) | Error::Utf8 => (),
-                        e => panic!("{e}")
-                    }
-                }
-            }
-            Err(_) => (),
-        });
-    }
-}
