--- src-ws-server/utils/mod.rs
@@ -1,3 +1,0 @@
-pub mod projects;
-pub mod extract;
-pub mod git_piping;
